backend/requirements.txt
Django==4.2.7
djangorestframework==3.15.2
django-filter==24.3
Pillow==10.1.0
psycopg2-binary==2.9.9
mysqlclient==2.2.7
django-redis==6.0.0
redis==6.2.0
channels==4.3.2
daphne==4.2.1